+++ app/src/main/java/vn/asiantech/way/ui/map/MapActivity.kt
@@ -1,13 +1,13 @@
 package vn.asiantech.way.ui.map
 
 import android.Manifest
-import android.os.Bundle
-import android.support.v4.app.Fragment
-import vn.asiantech.way.R
-import android.support.v4.app.ActivityCompat
 import android.content.pm.PackageManager
-import android.support.v4.content.ContextCompat
 import android.os.Build
+import android.os.Bundle
+import android.support.v4.app.ActivityCompat
+import android.support.v4.app.Fragment
+import android.support.v4.content.ContextCompat
+import vn.asiantech.way.R
 import vn.asiantech.way.ui.base.BaseActivity
 
 internal class MapActivity : BaseActivity() {
